requirements_manual.txt
django
djangorestframework
psycopg2
python-decouple
djangorestframework-simplejwt
django-sendgrid-v5
drf-yasg
django-cors-headers
google-api-python-client